tests/test_flux_analyzer_example.py
"""
Example test file for flux_analyzer module.
This demonstrates the structure and patterns for FluxMD tests.
"""

import pytest
import numpy as np
import pandas as pd
from unittest.mock import Mock, patch
import tempfile
import os

# These imports will need to be updated after restructuring
# from fluxmd.core.flux_analyzer import TrajectoryFluxAnalyzer


class TestTrajectoryFluxAnalyzer:
    """Test suite for TrajectoryFluxAnalyzer class"""
    
    @pytest.fixture
    def sample_interaction_data(self):
        """Create sample interaction data for testing"""
        data = {
            'frame': [0, 0, 1, 1, 2, 2],
            'residue_id': ['A:GLU:123', 'A:ASP:124', 'A:GLU:123', 'A:LYS:125', 'A:GLU:123', 'A:ASP:124'],
            'interaction_type': ['H-bond', 'Salt bridge', 'H-bond', 'Pi-cation', 'H-bond', 'Salt bridge'],
            'energy': [-3.5, -4.2, -3.3, -2.8, -3.6, -4.0],
            'distance': [2.8, 3.2, 2.9, 4.5, 2.7, 3.3]
        }
        return pd.DataFrame(data)
    
    @pytest.fixture
    def sample_trajectory_data(self):
        """Create sample trajectory data for testing"""
        n_frames = 100
        data = {
            'frame': np.arange(n_frames),
            'x': np.random.randn(n_frames) * 10,
            'y': np.random.randn(n_frames) * 10,
            'z': np.random.randn(n_frames) * 10,
        }
        return pd.DataFrame(data)
    
    @pytest.fixture
    def analyzer(self):
        """Create a TrajectoryFluxAnalyzer instance for testing"""
        # Once restructured:
        # return TrajectoryFluxAnalyzer()
        pass
    
    def test_bootstrap_analysis(self, analyzer, sample_interaction_data):
        """Test bootstrap statistical analysis"""
        # Test that bootstrap generates correct number of samples
        # Test that confidence intervals are properly calculated
        # Test that p-values are between 0 and 1
        pass
    
    def test_effect_size_calculation(self, analyzer, sample_interaction_data):
        """Test Cohen's d effect size calculation"""
        # Test with known values
        # Test edge cases (zero variance, etc.)
        pass
    
    def test_flux_calculation(self, analyzer, sample_interaction_data, sample_trajectory_data):
        """Test energy flux calculation"""
        # Test flux formula: Φᵢ = ⟨|Eᵢ|⟩ · Cᵢ · (1 + τᵢ)
        # Test directional consistency calculation
        # Test temporal fluctuation rate
        pass
    
    def test_empty_data_handling(self, analyzer):
        """Test handling of empty datasets"""
        empty_df = pd.DataFrame()
        # Should handle gracefully without crashing
        pass
    
    def test_output_file_generation(self, analyzer, sample_interaction_data):
        """Test that output files are correctly generated"""
        with tempfile.TemporaryDirectory() as tmpdir:
            # Run analysis
            # Check that expected files are created
            # Verify file contents
            pass
    
    @pytest.mark.parametrize("n_bootstrap", [100, 500, 1000])
    def test_bootstrap_iterations(self, analyzer, sample_interaction_data, n_bootstrap):
        """Test different numbers of bootstrap iterations"""
        # Verify that results stabilize with more iterations
        pass
    
    def test_interaction_filtering(self, analyzer, sample_interaction_data):
        """Test filtering of interactions by type"""
        # Test filtering by interaction type
        # Test energy thresholds
        # Test distance cutoffs
        pass


class TestStatisticalFunctions:
    """Test statistical utility functions"""
    
    def test_bootstrap_sampling(self):
        """Test bootstrap resampling implementation"""
        data = np.random.randn(100)
        # Test that samples are drawn with replacement
        # Test that sample size matches original
        pass
    
    def test_confidence_interval_calculation(self):
        """Test 95% confidence interval calculation"""
        data = np.random.randn(1000)
        # Test that CI contains the true mean ~95% of the time
        pass
    
    def test_p_value_calculation(self):
        """Test p-value calculation for flux significance"""
        # Test with known distributions
        # Test edge cases
        pass


class TestVisualization:
    """Test visualization components"""
    
    def test_heatmap_generation(self, sample_interaction_data):
        """Test that heatmaps are correctly generated"""
        # Test color mapping
        # Test axis labels
        # Test file output
        pass
    
    def test_colormap_normalization(self):
        """Test colormap normalization for flux values"""
        # Test with various flux ranges
        # Test log scaling option
        pass


@pytest.mark.integration
class TestIntegration:
    """Integration tests for complete flux analysis pipeline"""
    
    def test_full_analysis_pipeline(self):
        """Test complete analysis from raw data to results"""
        # Load test protein and ligand
        # Generate trajectory
        # Calculate interactions
        # Perform flux analysis
        # Verify outputs
        pass
    
    def test_error_propagation(self):
        """Test that errors are properly propagated through pipeline"""
        # Test with malformed input
        # Verify error messages are informative
        pass


if __name__ == "__main__":
    pytest.main([__file__, "-v"])