requirements-dev.txt
# Development dependencies for FluxMD

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0  # Parallel test execution
pytest-benchmark>=3.4.0  # Performance testing
pytest-mock>=3.10.0
coverage>=7.0.0

# Code quality
black>=23.0.0
flake8>=6.0.0
flake8-docstrings>=1.7.0
flake8-import-order>=0.18.0
pylint>=2.16.0
mypy>=1.0.0
ruff>=0.0.250  # Fast Python linter

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
sphinx-autodoc-typehints>=1.22.0
myst-parser>=0.19.0  # Markdown support for Sphinx
nbsphinx>=0.9.0  # Jupyter notebook support

# Development tools
ipython>=8.0.0
jupyter>=1.0.0
jupyterlab>=3.6.0
pre-commit>=3.0.0
twine>=4.0.0  # For PyPI uploads
build>=0.10.0  # For building packages

# Profiling and debugging
memory-profiler>=0.60.0
line-profiler>=4.0.0
py-spy>=0.3.0  # Sampling profiler
pdbpp>=0.10.0  # Enhanced Python debugger

# Type stubs
types-requests
types-PyYAML
pandas-stubs

# Security scanning
bandit>=1.7.0
safety>=2.3.0